refactor(TextLink): UpwardLinkをclient/配下へ移す
UpwardLinkはuseEnvironmentを利用するためclient componentが確定する。
TextLink・HelpLinkはServer Componentとして成立するため、
CLAUDE.mdのclient/ディレクトリ規約に沿って境界を分離する。

client/index.tsは作らず、client/components/index.tsが
'use client'を持つUpwardLinkのみをexportする構成のため、
バレルの平坦化による副作用importの混入は発生しない。
