+++ packages/smarthr-ui/src/components/TextLink/client/components/UpwardLink/index.ts
@@ -1,0 +1,1 @@
+export { UpwardLink } from './UpwardLink'
