+++ packages/smarthr-ui/src/components/TextLink/client/components/UpwardLink/stories/VRTUpwardLink.stories.tsx
@@ -1,4 +1,4 @@
-import { Stack } from '../../../Layout'
+import { Stack } from '../../../../../Layout'
 import { UpwardLink } from '../UpwardLink'
 
 import type { Meta, StoryObj } from '@storybook/react-webpack5'
