+++ packages/smarthr-ui/src/components/TextLink/index.ts
@@ -1,3 +1,3 @@
 export { TextLink } from './TextLink'
 export { HelpLink } from './HelpLink'
-export { UpwardLink } from './UpwardLink'
+export { UpwardLink } from './client/components'
